backend/requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
python-multipart
langchain
